Managinf Product Mudole for better and cleaner code also add Requests

+++ Modules/Product/database/migrations/2025_08_16_100352_create_products_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('price');
-            $table->integer('discount')->default(0)->nullable();
+            $table->unsignedBigInteger('price');
+            $table->unsignedTinyInteger('discount')->default(0)->nullable();
             $table->enum('availability_status', ['coming_soon', 'available', 'unavailable'])->default('available');
             $table->boolean('status')->default(true); // active/inactive
             $table->text('description')->nullable();
